add 'clock' script
This wraps the timecard program to provide a
pseudo-"distributed" timecard. We use a git repository to
push and pull changes from a central location, which should
notice any problems with multiple simultaneous updates
(though it will do nothing but just barf and let you fix up
the result yourself, so you do need to be at least a little
bit git-competent).