clock: avoid "git add -A"
For the most part, timecard repos will just have a timecard file in
them. But you may want to have other scripts tag along, or archive
invoiced cards in their own directory, etc. If you clock in or out while
an operation is in progress, that leads to very confusing results as
your in-progress state (including garbage temporary files!) is attached
to some random commit.

We know the name of the timecard file, since we just edited it! Just
commit that one file. We'll use git-commit's --only mode (this is the
default if paths are given), to make sure we don't likewise affect any
in-progress work in the index.

We do still have to "git add" the timecard file, since on the very first
clock in it won't yet be tracked.  It's a slight inefficiency in every
other case, but probably not worth worrying about (especially compared
to the fetch/push).