daily: de-dup comments
It's easy to clock in and out multiple times while working on some topic
(e.g., due to breaks or lunch in the middle). In the "daily" view all of
the comments for a single day are grouped, but you may end up with
duplicates, like:

  2022-11-01 3.25
   - foo
   - foo
   - something else
   - foo

Let's coalesce all of those into a single comment that we worked on
"foo".